+++ .travis.yml
@@ -12,7 +12,7 @@ install:
   - pip install -r requirements.txt
 
 script:
-  - pytest tests
+  - pytest tests/test_version1.py
 
 after_script:
   - coverage xml
